# be/storage/api-docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Laravel",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://speaks-alfred-technologies-gst.trycloudflare.com/api"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "security": [
        {
            "http": []  
        }
    ],
    "paths": {
...
        "/users/pending": {
            "get": {
                "operationId": "user.pendingUsers",
                "tags": ["User"],
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Terjadi kesalahan pada server. Silakan coba lagi."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Daftar user menunggu persetujuan berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "required": ["success", "message", "data"]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/users/aktif": {
            "get": {
                "operationId": "user.aktifUsers",
                "tags": ["User"],
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Terjadi kesalahan pada server. Silakan coba lagi."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Daftar user aktif berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "required": ["success", "message", "data"]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/users/ditolak": {
            "get": {
                "operationId": "user.ditolakUsers",
                "tags": ["User"],
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Terjadi kesalahan pada server. Silakan coba lagi."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Daftar user ditolak berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "required": ["success", "message", "data"]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/users/nonaktif": {
            "get": {
                "operationId": "user.nonaktifUsers",
                "tags": ["User"],
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Terjadi kesalahan pada server. Silakan coba lagi."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Daftar user non aktif berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/User"
                                            }
                                        }
                                    },
                                    "required": ["success", "message", "data"]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/users/{id}": {
            "put": {
                "operationId": "user.activate",
                "tags": ["User"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ActivateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Terjadi kesalahan pada server. Silakan coba lagi."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "Akun user berhasil diaktifkan dan ID dealer berhasil ditambahkan."
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "$ref": "#/components/schemas/StatusUser"
                                                },
                                                "email_verifikasi": {
                                                    "type": "boolean"
                                                },
                                                "id_dealer": {
                                                    "type": ["integer", "null"]
                                                },
                                                "dealer": {
                                                    "type": ["object", "null"],
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        },
                                                        "alamat": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": [
                                                        "id",
                                                        "name",
                                                        "alamat"
                                                    ]
                                                },
                                                "id_role": {
                                                    "type": "integer"
                                                },
                                                "role": {
                                                    "type": ["object", "null"],
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": ["id", "name"]
                                                },
                                                "id_atasan": {
                                                    "type": ["integer", "null"]
                                                },
                                                "atasan": {
                                                    "type": ["object", "null"],
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "type": "string"
                                                        }
                                                    },
                                                    "required": ["id", "name"]
                                                }
                                            },
                                            "required": [
                                                "id",
                                                "name",
                                                "email",
                                                "status",
                                                "email_verifikasi",
                                                "id_dealer",
                                                "dealer",
                                                "id_role",
                                                "role",
                                                "id_atasan",
                                                "atasan"
                                            ]
                                        }
                                    },
                                    "required": ["success", "message", "data"]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "const": "User tidak ditemukan."
                                        }
                                    },
                                    "required": ["success", "message"]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "ActivateUserRequest": {
                "type": "object",
                "properties": {
                    "id_dealer": {
                        "type": "integer"
                    },
                    "status": {
                        "$ref": "#/components/schemas/StatusUser"
                    }
                },
                "required": ["id_dealer"],
                "title": "ActivateUserRequest"
            },
...
            "StatusUser": {
                "type": "string",
                "enum": ["pending", "aktif", "ditolak", "nonaktif"],
                "title": "StatusUser"
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "email_verifikasi": {
                        "type": "boolean"
                    },
                    "status": {
                        "$ref": "#/components/schemas/StatusUser"
                    },
                    "foto": {
                        "type": ["string", "null"]
                    },
                    "created_at": {
                        "type": ["string", "null"],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": ["string", "null"],
                        "format": "date-time"
                    },
                    "id_dealer": {
                        "type": ["integer", "null"]
                    },
                    "id_atasan": {
                        "type": ["integer", "null"]
                    },
                    "id_role": {
                        "type": "integer"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "email",
                    "email_verifikasi",
                    "status",
                    "foto",
                    "created_at",
                    "updated_at",
                    "id_dealer",
                    "id_atasan",
                    "id_role"
                ],
                "title": "User"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": ["message", "errors"]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": ["message"]
                        }
                    }
                }
            }
        }
    }
}